add metagenome submitter class
